Added: KQOAuthReplyServer.
Included a HTTP server that will listen for the response from the
authorization server and parse the query params from the request. We will
use this for easy encapsulation to retrieve the verifier.

Next we will integrate this server interaction to a convenient wrapper
that will take care of all the flow - and the user can then immediately
start sending authorized requests.