Added requestTimeout() slot to handle requestTimedout() signals emitted from a KQOAuthRequest.
Use requestMap to map QNetworkReplies to KQOAuthRequest.  This is used in timeout to initiate
an abort() on the QNetworkReply.

Use requestMap to map QNetworkReplies in the slots that are connected to QNetworkReply finished() signals.
This allows the timeout and error handling code to run correctly if multiple KQOAuthRequest are active
at the same time.